fiscal-facil-3.1/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
xmltodict
requests
python-multipart
streamlit
pandas